Leak the JavascriptEnvironment on exit
This is to work around the bug that V8 would be waiting for background
tasks to finish on exit, while somehow it waits forever in Electron, more
about this can be found at https://github.com/atom/electron/issues/4767.

On the other handle there is actually no need to gracefully shutdown V8
on exit in the main process, we already ensured all necessary resources get
cleaned up, and it would make quitting faster.